+++ src/analyze/src/main.js
@@ -1,21 +1,22 @@
 import Vue from 'vue';
 import { library } from '@fortawesome/fontawesome-svg-core';
 import {
-  faChartBar,
-  faChartLine,
-  faChartArea,
-  faChartPie,
-  faGlobeAmericas,
-  faCertificate,
-  faDotCircle,
-  faCaretDown,
-  faCaretUp,
   faAngleDown,
   faAngleUp,
-  faSearch,
-  faHashtag,
-  faExclamationTriangle,
   faArrowRight,
+  faCaretDown,
+  faCaretUp,
+  faCertificate,
+  faChartArea,
+  faChartBar,
+  faChartLine,
+  faChartPie,
+  faDotCircle,
+  faExclamationTriangle,
+  faGlobeAmericas,
+  faHashtag,
+  faSearch,
+  faUser,
 } from '@fortawesome/free-solid-svg-icons';
 import {
   FontAwesomeIcon,
@@ -30,21 +31,22 @@ import Auth from './auth';
 
 Vue.config.productionTip = false;
 
-library.add(faChartBar);
-library.add(faChartLine);
-library.add(faChartArea);
-library.add(faChartPie);
-library.add(faGlobeAmericas);
-library.add(faCertificate);
-library.add(faDotCircle);
-library.add(faCaretDown);
-library.add(faCaretUp);
 library.add(faAngleDown);
 library.add(faAngleUp);
-library.add(faSearch);
-library.add(faHashtag);
-library.add(faExclamationTriangle);
 library.add(faArrowRight);
+library.add(faCaretDown);
+library.add(faCaretUp);
+library.add(faCertificate);
+library.add(faChartArea);
+library.add(faChartBar);
+library.add(faChartLine);
+library.add(faChartPie);
+library.add(faDotCircle);
+library.add(faExclamationTriangle);
+library.add(faGlobeAmericas);
+library.add(faHashtag);
+library.add(faSearch);
+library.add(faUser);
 
 Vue.component('font-awesome-icon', FontAwesomeIcon);
 Vue.component('font-awesome-layers', FontAwesomeLayers);